Change model instances to inherit from EventListener
Model associations can now have options. The first implemented
options is autoFetch that retrieves the associations before
returning the records you want.

**BE CAREFULL FOR CIRCULAR OR VERY DEEP ASSOCIATIONS**

To achieve this, I changed Model.hasOne() to support arguments
in any order, a string being the association name, an (optional)
Model (if not the same) and an object with options. Example:

Person.hasOne("Mother", { autoFetch: true });

This is still experimental but feel free to post any issues or
suggestions.